server/requirements.txt
Flask
gunicorn
google-cloud-storage
firebase-admin
numpy
pillow
